requirements.txt
flask==3.0.0
pyTelegramBotAPI==4.14.0
openai==1.51.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
httpx==0.24.1

